Do not share entire FrameRenderer with layout, only current frame
